Default to XDG dir for configuration
This patch changes the default configuration directory, internally
referred to as 'ApplicationDir', to comply with XDG's Base Directory
Specification.

Most Unix tools already operate on the specification. By adhering it,
Nchat becomes more intuitive and integrated with the system.

More information about the spec can be found at:
    https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

The '--confdir' flag doesn't cover this scenario as it's directed at
customization at invokation and not at defaults.

Signed-off-by: Carlos Torres <torr.ca@mailgw.com>